User profile page
Moved sessions into base controller

Updated get_user_profile

Created new template "user_profile.html"

Updated users.py

Modified views.py
