+++ ntss_www/ntss/views/users.py
@@ -14,5 +14,8 @@ class UserViews(Views):
         Load the the profile for a user
         """
         self.set_template('user_profile.html')
+        self.template_vars['states'] = self.US_STATES
+        self.template_vars['roles'] = self.ROLES
         self.template_vars['user'] = user_info
+        print(user_info)
         return self.template.render(self.template_vars)
