+++ ntss_www/ntss/views/views.py
@@ -3,7 +3,7 @@ This is the base module for Views
 All Views should inherit from this class
 """
 from jinja2 import Environment, FileSystemLoader, select_autoescape
-from ntss.config.constants import WWW_PATH
+from ntss.config.constants import WWW_PATH, US_STATES, ROLES
 
 
 class Views:
@@ -15,7 +15,9 @@ class Views:
     _templates_path = f'{WWW_PATH}ntss/templates'
     _env = Environment()
     __platformName = 'NTSS'
-
+    US_STATES = US_STATES
+    ROLES = ROLES
+    
     def __init__(self):
         self.template_vars = {
             'platformName': self.__platformName,
